remi_ai/server/python_service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
opencv-python>=4.9.0
insightface>=0.7.3
onnxruntime>=1.17.0
faiss-cpu>=1.8.0
